sdk: regenerate against the latest api
- to vms.add() added [action.vm.initialization.cloud-init]
- to NIC added OnBoot/BootProtocol properties
- to VersionCaps added a list of supported payload-encodings
- to Step added externalType
- to NIC added vnicProfile and bootProtocol
- to CPU added architecture
- to VnicProfilePermission added delete() method
- to Disk added readOnly
- to VMs.add() added [vm.cpu.architecture], [action.vm.initialization.cloud_init.*] arguments
- to Templates.add() added [template.cpu.architecture], [action.template.initialization.cloud_init.*] arguments
- to UserRoles.add() added permit.id|name arguments
- at VMSnapshot removed preview/undo/commit methods
- to DataCenterClusterGlusterVolumeGlusterBricks added activate/stopmigrate/migrate actions
- to NetworkVnicProfile added Permissions sub-collection
- to Cluster added [cluster.cpu.architecture]
- to DataCenter added Networks sub-collection
- to ClusterGlusterVolumeGlusterBricks added activate method
- to ClusterGlusterVolume added stoprebalance method
- to entry-point API added Permissions collection (for managing system-permissions)

+++ ovirt-engine-sdk-java/src/main/java/org/ovirt/engine/sdk/entities/GlusterClient.java
@@ -39,8 +39,8 @@ import javax.xml.bind.annotation.XmlType;
  *       &lt;sequence>
  *         &lt;element name="host_name" type="{http://www.w3.org/2001/XMLSchema}string" minOccurs="0"/>
  *         &lt;element name="client_port" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
- *         &lt;element name="bytes_read" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
- *         &lt;element name="bytes_written" type="{http://www.w3.org/2001/XMLSchema}int" minOccurs="0"/>
+ *         &lt;element name="bytes_read" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/>
+ *         &lt;element name="bytes_written" type="{http://www.w3.org/2001/XMLSchema}long" minOccurs="0"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
@@ -63,9 +63,9 @@ public class GlusterClient {
     @XmlElement(name = "client_port")
     protected Integer clientPort;
     @XmlElement(name = "bytes_read")
-    protected Integer bytesRead;
+    protected Long bytesRead;
     @XmlElement(name = "bytes_written")
-    protected Integer bytesWritten;
+    protected Long bytesWritten;
 
     /**
      * Gets the value of the hostName property.
@@ -128,10 +128,10 @@ public class GlusterClient {
      *
      * @return
      *     possible object is
-     *     {@link Integer }
+     *     {@link Long }
      *
      */
-    public Integer getBytesRead() {
+    public Long getBytesRead() {
         return bytesRead;
     }
 
@@ -140,10 +140,10 @@ public class GlusterClient {
      *
      * @param value
      *     allowed object is
-     *     {@link Integer }
+     *     {@link Long }
      *
      */
-    public void setBytesRead(Integer value) {
+    public void setBytesRead(Long value) {
         this.bytesRead = value;
     }
 
@@ -156,10 +156,10 @@ public class GlusterClient {
      *
      * @return
      *     possible object is
-     *     {@link Integer }
+     *     {@link Long }
      *
      */
-    public Integer getBytesWritten() {
+    public Long getBytesWritten() {
         return bytesWritten;
     }
 
@@ -168,10 +168,10 @@ public class GlusterClient {
      *
      * @param value
      *     allowed object is
-     *     {@link Integer }
+     *     {@link Long }
      *
      */
-    public void setBytesWritten(Integer value) {
+    public void setBytesWritten(Long value) {
         this.bytesWritten = value;
     }
